Ajout de l'authentification JWT et d'une sécurité qui rend impossible la suppression/ la modification d'une donnée qui ne nous appartient pas

+++ src/routes/box/delete.js
@@ -2,7 +2,19 @@ const { prisma } = require("../../prisma")
 
 async function deleteBox(req, res) {
   try {
+    const id_user = req.auth.id
     const id = req.body.id
+
+    //vérification que c'est le bon utilisateur
+    const box = await prisma.box.findUnique({
+      where: {
+        id: id
+      }
+    })
+    if (id_user != box.id_user) {
+      return res.status(403).send("Vous n'êtes pas autorisé à faire ceci")
+    }
+
     //on supprime la boite de la bdd
     await prisma.box.delete({
       where: {
